requirements-dev.txt
-e .
absl-py==0.7.1
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
decorator==4.3.2
docopt==0.6.2
ipython==7.2.0
ipython-genutils==0.2.0
more-itertools==7.0.0
numpy==1.16.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.9.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
py==1.8.0
Pygments==2.3.1
pygtrie==2.3
pytest==4.4.0
scipy==1.2.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
